Moved UserDataValidation, fixed it's usings
