+++ src/backend/AuthService/BusinessLogic/GetCurrentUser/GetCurrentUserService.cs
@@ -27,10 +27,10 @@ namespace AuthService.BusinessLogic.GetCurrentUser
 
         public async Task<User?> GetCurrentUser(CurrentUserInfo auth0UserInfo)
         {
-            var userFromDatabase = await _appDbContext.Auth0Users.Where(c => c.Email == auth0UserInfo.Email).Select(c => c.User).SingleOrDefaultAsync();
+            var userFromDatabase = await _appDbContext.Auth0Users.Where(c => c.Identifier == auth0UserInfo.Identifier).Select(c => c.User).SingleOrDefaultAsync();
             if (userFromDatabase == null)
             {
-                _logger.LogError("UserFromDatabase is null");
+                _logger.LogError($"User with email {auth0UserInfo.Email} was not found in table Auth0Users");
                 throw new HttpError500Exception("Sorry we couldn't find your user in database");
             }
