+++ src/backend/AuthService/BusinessLogic/PostLogin/PostLoginAction.cs
@@ -1,4 +1,5 @@
-﻿using Core.Auth0;
+﻿using AuthService.Services;
+using Core.Auth0;
 using Core.Exceptions;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
@@ -28,13 +29,7 @@ namespace AuthService.BusinessLogic.PostLogin
         public async Task<ActionResult> Execute()
         {
             var auth0UserInfo = await _currentUserService.GetUserInfo();
-            var isUserInfoValid = _userDataValidationService.Check(auth0UserInfo);
-
-            if (!isUserInfoValid)
-            {
-                _logger.LogError("Auth0UserInfo in PostLoginAction didn't pass validation as it has nulls");
-                throw new HttpError500Exception("Your Auth0User data is invalid");
-            }
+            _userDataValidationService.Check(auth0UserInfo);
 
             var user = await _userService.CreateUserAndAddToDatabase(auth0UserInfo);
